implemented all requested changes
