+++ src/app/components/forms/add/add-money-to-piggy/add-money-to-piggy.component.ts
@@ -1,9 +1,9 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { Account } from 'src/app/models/Account.model';
-import { PiggyBank } from 'src/app/models/PiggyBank.model';
-import { AccountService } from 'src/app/services/accountService/account.service';
-import { PiggyBankService } from 'src/app/services/piggyBankService/piggy-bank.service';
+import { Account } from 'src/app/models/account.model';
+import { PiggyBank } from 'src/app/models/piggyBank.model';
+import { AccountService } from 'src/app/services/account.service';
+import { PiggyBankService } from 'src/app/services/piggy-bank.service';
 
 @Component({
   selector: 'app-add-money-to-piggy',
