+++ src/app/components/user-actions/user-actions.component.ts
@@ -1,10 +1,10 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
 import { filter, switchMap } from 'rxjs/operators';
-import { Account } from 'src/app/models/Account.model';
-import { PiggyBank } from 'src/app/models/PiggyBank.model';
-import { AccountService } from 'src/app/services/accountService/account.service';
-import { PiggyBankService } from 'src/app/services/piggyBankService/piggy-bank.service';
+import { Account } from 'src/app/models/account.model';
+import { PiggyBank } from 'src/app/models/piggyBank.model';
+import { AccountService } from 'src/app/services/account.service';
+import { PiggyBankService } from 'src/app/services/piggy-bank.service';
 
 
 @Component({
