+++ src/app/constants/constants.ts
@@ -1,0 +1,3 @@
+export class Constants {
+  static apiUrl = 'http://localhost:8080';
+}
