+++ src/app/models/Account.model.ts
@@ -1,4 +1,4 @@
-import { User } from "./User.model";
+import { User } from "./user.model";
 
 export interface Account {
   id?: number;
