+++ src/app/redux/actions/user.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { User } from 'src/app/models/User.model';
+import { User } from 'src/app/models/user.model';
 
 export const setUser = createAction('[User] Set User', props<{ user: User }>());
 export const getUserInfo = createAction('[User] Get User Info');
